Add display order to maltAddition CoreData object
This conflates UI with the model a little bit, but it would be
too much work to build some separate UI sorting component.
This will eventually be used to persist the ordering of malts
across sessions. It also might simplify some of the table view
code a bit (not having to separate malt lists).