Simplify OBIngredientDashboardController
This controller previously tried to allow itself to be reused
for both displaying a table of hops and a table of malts.
These two ingredient types are two different to lend themselves
to be reused. This complicated the code more than it helped.

The next commit will rename the view controller to make it clear
that it supports displaying malts rather than general ingredients.