Gravity and IBU gauges work again
After some previous refactoring of the malt and hop view controllers,
the gauge readings no longer updated.  This was fixed by implementing
KVO on all hop and malt variables. When variables change, a notification
is triggered for objects observing the OBRecipe's calculated values
(such as IBUs). The controllers that own the gauges now register to
observe these variables and can update the gauges when changes are made.