Recipes and their ingredients are being saved
One bug introduced: ingredient gauges no longer get updated
when pickers are selected. This is possibly due to KVO
registration not getting setup for malts/hops that already
exist in a recipe... not entirely sure, though.

Don't merge back into the main branch until this is fixed.
At that point, squash the commits.

Saving for the night...