Refactor: OBDrawerTableViewDelegate now owns cell 'move' and 'delete' logic
This commit will facilitate work on issue #14

Previously this was implemented in both OBMaltAdditionTableViewDelegate.m
and OBHopAdditionTableViewDelegate.m.  The code was largely duplicated.
In order to do this, a new protocol was formed: OBIngredientAddition.
This protocol allows ingredients to remove themselves from recipes.  This
is a bit awkward, but causes no problems at the moment.