Get rid of picker observers. Use KVO instead.
Hop, Malt, and BatchSize controllers now use KVO to determine when
to reload the data in the table views.

This introduces a bug where the segmented control will get reset
whenever the picker changes. I'll address this in one of the
next commits.

This is cutting out some of the spaghetti from the picker update
logic. Exciting!