Massive simlification for hop and malt controllers
Picker updates are far more simple.
Moved OBMultiPIckerTableViewCell logic into a self contained
component: OBMultiPickerView.

The MultiPickerView takes care of switching between pickers
rather than the TableViewDelegates.  This makes way more sense,
and the code is substantially simpler because of it.

This commit breaks OBBatchSizeTableViewController, but
it is time to go enjoy the sunshine.  I'll fix it in the next
commit.