Bugfix: OBMultiPickerView would reset when a picker value was chosen
Syptom:
In the OBHopAdditionViewController or OBMaltAdditionTableViewController,
when a drawer was open and a picker value was chosen, the picker would
get reset to the first picker in the OBMultiPickerView.

RCA:
When a picker value was chosen, tableView reload was being called
on the table. A new cell would be created and it wouldn't maintain
the previous state.

Fix:
Now the malt and hop table view delegates maintain the index of the
currently selected picker.