MaltAddition and HopAddition inherit from Malt and Hops
I debated whether this was the right thing to do for a while.
On the surface a MaltAddition is not a "type of" Malt, so
inheritance doesn't make sense from an intuitive perspective.
However, it does accomplish the goal of sharing some code.
It also makes the data model a bit more clear IMO.