Stop pre-generating the starter DB. Just load the data from a CSV.
Since the dataset is so small, the performance benefit of doing this
was negligible. More importantly, running an OSX target from an iOS
target doesn't seem to be supported (I ran into problems with
creating an archive and also command line builds were breaking).
The only other option was building hte database manually, which is
unacceptable IMO.