Go back to using the treating the OBSettings as a singleton
I'm a bit disappointed, but a singleton does seem to be
the best thing here.  I don't like the recipe object deciding
which IBU formula to use, but it really does simplify the code.

Most importantly it allows treating the IBU property as a
key-value-coding complaint property, which will allow decoupling
the OBGaugeViewController from OBRecipe. This makes
OBGaugeViewController reusable in the myriad calculation screens.