Break up the OBGaugeViewController into two classes
OBNumericGaugeViewController
OBColorGaugeViewController

Also, make these classes more generic. They're no longer
coupled to OBRecipe.  They simply take a 'target' (of type 'id')
and a key that will be displayed on the gauge.

This makes it possible to reuse the gauge view for the
calculation controllers which do not use OBRecipe.