Bugfix: Slow launches.
Symptom:
Tap the app icon and the OS seems to freeze for 2-3 seconds
before actually launching the app.

RCA:
The Main.storyboard was failing to be loaded as a launch image.

Fix:
Create a new LaunchScreen.storyboard.  Also add some grey images
for older OSs that can't use the LaunchScreen.