Add a search bar to the Malt Finder View
Removed the segmented control that allowed switching
between extracts, malts, and sugars. In the HomeBrewTalk
forum this seemed to cause people to thing various sugars
were not available. I think the search feature will potentially
be a faster way to get to the correct thing anyways.

This also keeps the design and code simple.